requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-geojson==2.11.0
django-leaflet==0.24.0
django-octicons==1.0.2
django-social-share==1.3.0
jsonfield==2.0.2
olefile==0.46
Pillow==5.3.0
pytz==2018.7
six==1.11.0
